sitemap_pro/OCRappBackupFile/CoreOCR&ScrapingBackup/251220_NewOCR_web00/requirements.txt
click>=8.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.5.0
numpy>=1.21.0
pdf2image>=1.16.0
google-cloud-vision>=3.0.0

